+++ packages/astro/src/compiler/codegen/index.ts
@@ -1,4 +1,4 @@
-import type { Ast, Script, Style, TemplateNode } from 'astro-parser';
+import type { Ast, Script, Style, TemplateNode } from '@astrojs/parser';
 import type { CompileOptions } from '../../@types/compiler';
 import type { AstroConfig, AstroMarkdownOptions, TransformResult, ComponentInfo, Components } from '../../@types/astro';
 import type { ImportDeclaration, ExportNamedDeclaration, VariableDeclarator, Identifier, ImportDefaultSpecifier } from '@babel/types';
@@ -7,7 +7,7 @@ import 'source-map-support/register.js';
 import eslexer from 'es-module-lexer';
 import esbuild from 'esbuild';
 import path from 'path';
-import { parse } from 'astro-parser';
+import { parse } from '@astrojs/parser';
 import { walk, asyncWalk } from 'estree-walker';
 import _babelGenerator from '@babel/generator';
 import babelParser from '@babel/parser';