+++ packages/astro/src/logger.ts
@@ -1,5 +1,5 @@
 import 'source-map-support/register.js';
-import type { CompileError } from 'astro-parser';
+import type { CompileError } from '@astrojs/parser';
 import { bold, blue, red, grey, underline } from 'kleur/colors';
 import { Writable } from 'stream';
 import { format as utilFormat } from 'util';