+++ packages/astro/test/fixtures/preact-component/src/components/Class.jsx
@@ -1,0 +1,7 @@
+import { h, Component } from 'preact';
+
+export default class extends Component {
+  render() {
+    return <div id="class-component"></div>
+  }
+}