+++ packages/astro/test/fixtures/preact-component/src/components/Function.jsx
@@ -1,0 +1,5 @@
+import { h, Component } from 'preact';
+
+export default function() {
+  return <div id="fn-component"></div>;
+}