+++ packages/astro/test/fixtures/react-component/src/components/ForgotImport.jsx
@@ -1,0 +1,5 @@
+
+
+export default function ({}) {
+  return <h2>oops</h2>;
+}